fix bug in packing/unpacking of files
*** Unpacking bug
When handling file data while unpacking, sometimes it would read more
data than the file size from the BAG archive file. This is because, I
was using `BufReader::read` method, which reads variable amount of data.
So when there was X amount of bytes remaining it could read more than X
bytes of data, thereby reading more bytes after the file data in the BAG
archive file. As a result, in the next iteration it would parse the
header correctly and would, essentially, end up corrupting the rest of
the file.

This is fixed by reading exact amount of data as the file size, from the
archive file, in chunks. It now calculates correctly to always read the
exact amount of file data and not overshoot. This also moves the
BufReader call to `read_exact` instead of `read`.

*** Packing bug
While reading source files > 8KB, it would read the file in chunks of
8KB and using the `BufReader::read` method. As it read variable amount
of data, it would not always read full 8KB data even though the buffer
size is 8KB. As a result the buffer could have lot of initial data (0s)
which is not part of the original file data. To circumvent this I would
filter out the buffer of only non-zero bytes. But this is wrong! Files
could have legitimate data of contigous zero bytes. So droppping them
would mean actually corrupting the data in the source files!

This is fixed by using the `bytes_read` output of the `BufReader::read`
call, which tells us how many bytes were read, and then only saving
those bytes from our 8KB buffer into the writer file. This is done by
simply slicing the 8KB buffer upto the `bytes_read` -

```rust
let buffer = [0u8; 8192];
loop {
     let bytes_read = reader.read(&buffer)?;
     writer.write(&buffer[..bytes_read])?;
}
```