Change the trait to have `pack_header`
Change the packer interface to have `pack_header` instead of
`pack_file`. For now, packing of the actual file data is same across all
backends, so instead of duplicating it in all backends, we do it in the
`pack` function now. And instead the PackerBackend only packs the header
via the new `pack_header` function.